# Add ua_list::STATIC_AGENTS and use that in ua::spoof_ua()