--- a/ua_generator/src/ua.rs
+++ b/ua_generator/src/ua.rs
@@ -1,9 +1,9 @@
-pub use crate::ua_list::agents;
+pub use crate::ua_list::STATIC_AGENTS;
 use fastrand;
 
 /// Get a random UA from the most popular list pre-compiled generated from <https://apilayer.com/marketplace/user_agent-api#documentation-tab>.
 pub fn spoof_ua() -> &'static str {
-    let agent_list = agents(); // list of user agents to choose from.
+    let agent_list = STATIC_AGENTS; // list of user agents to choose from.
     let i = fastrand::usize(..agent_list.len());
 
     agent_list[i]