This avoids a temp allocation every time `spoof_ua()` is called and, not that it's exactly critical, improves performance

Before:
```
generate/personal/basic: generate 1000 samples                                                                              
                        time:   [11.945 ns 11.994 ns 12.051 ns]
Found 62 outliers among 1000 measurements (6.20%)
  7 (0.70%) high mild
  55 (5.50%) high severe
```

After:
```
generate/personal/basic: generate 1000 samples                                                                              
                        time:   [4.3748 ns 4.3931 ns 4.4100 ns]
                        change: [-65.797% -65.512% -65.232%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 1000 measurements (0.20%)
  2 (0.20%) high mild
```

I kept the original `ua_list::agents()` function to maintain backward compatibility.

I also split the list out to one per line just to make it easier to read, but I can revert that if it's not desirable.