@j-mendez 2022-12-30
I forgot to mention we need the change applied in `build.rs`. Pushed the changes here https://github.com/a11ywatch/ua_generator/commit/0f0d066bc989d757b2868a782bf6e06cb0555fc9, released in `0.3.5`.